fix all the test warnings
