+++ crates/ndc-sqlserver/tests/explain_tests.rs
@@ -1,6 +1,6 @@
-use crate::common::{is_contained_in_lines, run_explain};
+use crate::common::helpers::{is_contained_in_lines, run_explain};
 
-mod common;
+pub mod common;
 
 #[tokio::test]
 async fn select_by_pk() {
