+++ crates/ndc-sqlserver/tests/query_tests.rs
@@ -1,7 +1,9 @@
-mod common;
+pub mod common;
 
+#[cfg(test)]
 mod basic {
-    use super::common::run_query;
+    use super::common::helpers::*;
+
     #[tokio::test]
     async fn select_by_pk() {
         let result = run_query("select_by_pk").await;
@@ -22,7 +24,7 @@ mod basic {
 }
 
 mod predicates {
-    use super::common::run_query;
+    use super::common::helpers::*;
 
     #[tokio::test]
     async fn select_where_name_like() {
@@ -146,7 +148,7 @@ mod predicates {
 }
 
 mod sorting {
-    use super::common::run_query;
+    use super::common::helpers::*;
 
     #[tokio::test]
     async fn select_order_by_name() {
@@ -210,7 +212,7 @@ mod sorting {
 }
 
 mod aggregation {
-    use super::common::run_query;
+    use super::common::helpers::*;
 
     #[tokio::test]
     async fn aggregate_count_albums() {
@@ -238,7 +240,7 @@ mod aggregation {
 }
 
 mod relationships {
-    use super::common::run_query;
+    use super::common::helpers::*;
     #[tokio::test]
     async fn select_album_object_relationship_to_artist() {
         let result = run_query("select_album_object_relationship_to_artist").await;
@@ -278,7 +280,7 @@ mod relationships {
 
 #[cfg(test)]
 mod native_queries {
-    use super::common::run_query;
+    use super::common::helpers::*;
 
     #[tokio::test]
     async fn select_artist() {
@@ -321,7 +323,7 @@ mod native_queries {
 
 #[cfg(test)]
 mod types {
-    use super::common::run_query;
+    use super::common::helpers::*;
 
     #[tokio::test]
     async fn select_value_types() {
