+++ crates/ndc-sqlserver/tests/schema_tests.rs
@@ -2,6 +2,6 @@ pub mod common;
 
 #[tokio::test]
 async fn get_schema() {
-    let result = common::get_schema().await;
+    let result = common::helpers::get_schema().await;
     insta::assert_json_snapshot!(result);
 }
